Change some +stringWithCString and -initWithCString:encoding: to modern Objective-C syntax. Change the encoding of a couple of info.plists to UTF-8, otherwise Xcode was complaining. Removing some extra line breaks on OS X code. Other minor changes.
git-svn-id: https://pcsxr.svn.codeplex.com/svn/pcsxr@85867 e17a0e51-4ae3-4d35-97c3-1a29b211df97

+++ plugins/dfxvideo/soft.c
@@ -1213,7 +1213,7 @@ static __inline int shl10idiv(int x, int y)
 {
  __int64 bi=x;
  bi<<=10;
- return bi/y;
+ return (int)(bi/y);
 }
 
 #endif
